pyproject.toml
[project]
name = "viralens"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "elevenlabs>=2.28.0",
    "fal-client>=0.11.0",
    "ipykernel>=7.1.0",
    "langchain[google-genai]>=1.2.0",
    "langgraph>=1.0.5",
    "python-dotenv>=1.2.1",
    "requests>=2.31.0",
    "tavily-python>=0.5.0",
]
